challs/bjorn_barcode/src/requirements.txt
code128
Pillow